+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/payum/command_handlers.xml
@@ -22,8 +22,8 @@
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
             <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
             <argument type="service" id="sylius.payum.payment_request.factory.payum_token" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessorInterface" />
+            <argument type="service" id="sylius.payum.payment_request.processor.request" />
+            <argument type="service" id="sylius.payum.payment_request.processor.after_token_request" />
         </service>
 
         <service id="sylius.payment_request.payum.capture" parent="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\TokenPaymentRequestHandler">
@@ -39,7 +39,7 @@
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\ModelPaymentRequestHandler" abstract="true">
             <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
             <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" />
+            <argument type="service" id="sylius.payum.payment_request.processor.request" />
         </service>
 
         <service id="sylius.payment_request.payum.status" parent="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\ModelPaymentRequestHandler">
