--- src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/payum/processors.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
->
-    <services>
-        <defaults public="true" />
-
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessorInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessor">
-            <argument type="service" id="sylius.factory.payment_request" />
-            <argument type="service" id="sylius.repository.payment_request" />
-            <argument type="service" id="sylius.announcer.payment_request" />
-        </service>
-
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessor">
-            <argument type="service" id="payum" />
-            <argument type="service" id="sylius.payum.context.payment_request" />
-        </service>
-
-    </services>
-</container>
