+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/processor.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.payum.payment_request.processor.after_token_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\Processor\AfterTokenRequestProcessor">
+            <argument type="service" id="sylius.factory.payment_request" />
+            <argument type="service" id="sylius.repository.payment_request" />
+            <argument type="service" id="sylius.announcer.payment_request" />
+        </service>
+        <service id="Sylius\Bundle\PayumBundle\PaymentRequest\Processor\AfterTokenRequestProcessorInterface" alias="sylius.payum.payment_request.processor.after_token_request" />
+
+        <service id="sylius.payum.payment_request.processor.request" class="Sylius\Bundle\PayumBundle\PaymentRequest\Processor\RequestProcessor">
+            <argument type="service" id="payum" />
+            <argument type="service" id="sylius.payum.context.payment_request" />
+        </service>
+        <service id="Sylius\Bundle\PayumBundle\PaymentRequest\Processor\RequestProcessorInterface" alias="sylius.payum.payment_request.processor.request" />
+    </services>
+</container>
